backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
hypothesis==6.92.1
pytest==7.4.3
reportlab==4.0.7
